feat: added governance v3 to proposal detials page
