+++ src/components/primitives/Link.tsx
@@ -11,7 +11,7 @@ const Anchor = styled('a')({});
 
 interface NextLinkComposedProps
   extends Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, 'href'>,
-    Omit<NextLinkProps, 'href' | 'as'> {
+  Omit<NextLinkProps, 'href' | 'as'> {
   to: NextLinkProps['href'];
   linkAs?: NextLinkProps['as'];
   href?: NextLinkProps['href'];
@@ -121,7 +121,7 @@ export const ROUTES = {
   faucet: '/faucet',
   migrationTool: '/v3-migration',
   prerenderedProposal: (proposalId: number) => `/governance/proposal/${proposalId}`,
-  dynamicRenderedProposal: (proposalId: number) => `/governance/proposal?proposalId=${proposalId}`,
+  dynamicRenderedProposal: (proposalId: number) => `/governance/proposal/${proposalId}`,
   reserveOverview: (underlyingAsset: string, marketName: CustomMarket) =>
     `/reserve-overview/?underlyingAsset=${underlyingAsset}&marketName=${marketName}`,
   history: '/history',
