+++ src/hooks/governance/useProposalVotes.ts
@@ -2,12 +2,19 @@ import { normalizeBN } from '@aave/math-utils';
 import { useQuery } from '@tanstack/react-query';
 import request, { gql } from 'graphql-request';
 
-type ProposalVote = {
+export type ProposalVote = {
   proposalId: string;
   support: boolean;
   voter: string;
   votingPower: string;
 };
+
+export interface ProposalVotes {
+  yaeVotes: ProposalVote[],
+  nayVotes: ProposalVote[],
+  combinedVotes: ProposalVote[],
+  isFetching: boolean,
+}
 
 const VOTING_MACHING_SUBGRAPH_URL =
   'https://api.goldsky.com/api/public/project_clk74pd7lueg738tw9sjh79d6/subgraphs/votingmachine-sepolia/v1/gn';
@@ -42,12 +49,17 @@ export const useProposalVotesQuery = ({ proposalId }: { proposalId: string }) =>
   });
 };
 
-export const useProposalVotes = ({ proposalId }: { proposalId: string }) => {
+const sortByVotingPower = (a: ProposalVote, b: ProposalVote) => {
+  return a.votingPower < b.votingPower ? 1 : a.votingPower > b.votingPower ? -1 : 0;
+};
+
+export const useProposalVotes = ({ proposalId }: { proposalId: string }): ProposalVotes => {
   const { data, isFetching } = useProposalVotesQuery({ proposalId });
 
   return {
-    yaeVotes: data?.filter((vote) => vote.support === true) || [],
-    nayVotes: data?.filter((vote) => vote.support === false) || [],
+    yaeVotes: data?.filter((vote) => vote.support === true).sort(sortByVotingPower) || [],
+    nayVotes: data?.filter((vote) => vote.support === false).sort(sortByVotingPower) || [],
+    combinedVotes: data?.sort(sortByVotingPower) || [],
     isFetching,
   };
 };
