+++ src/modules/governance/proposal/VoteInfo.tsx
@@ -1,35 +1,52 @@
-import { ProposalState } from '@aave/contract-helpers';
+import { ProposalState, ProposalV3State } from '@aave/contract-helpers';
 import { Trans } from '@lingui/macro';
-import { Button, Typography } from '@mui/material';
+import { Button, Paper, Typography } from '@mui/material';
 import { FormattedNumber } from 'src/components/primitives/FormattedNumber';
 import { Row } from 'src/components/primitives/Row';
 import { Warning } from 'src/components/primitives/Warning';
 import { ConnectWalletButton } from 'src/components/WalletConnection/ConnectWalletButton';
+import { EnhancedProposal } from 'src/hooks/governance/useProposal';
 import { useVoteOnProposal } from 'src/hooks/governance/useVoteOnProposal';
 import { useVotingPowerAt } from 'src/hooks/governance/useVotingPowerAt';
 import { useModalContext } from 'src/hooks/useModal';
 import { CustomProposalType } from 'src/static-build/proposal';
 import { useRootStore } from 'src/store/root';
 
-export function VoteInfo({ id, state, strategy, startBlock }: CustomProposalType) {
+interface VoteInfoProps {
+  proposal: EnhancedProposal;
+}
+
+export function VoteInfo({ proposal }: VoteInfoProps) {
   const { openGovVote } = useModalContext();
   const user = useRootStore((state) => state.account);
   const currentMarketData = useRootStore((store) => store.currentMarketData);
 
-  const { data: voteOnProposal } = useVoteOnProposal(currentMarketData, id);
-  const { data: powerAtProposalStart } = useVotingPowerAt(currentMarketData, strategy, startBlock);
+  // const { data: voteOnProposal } = useVoteOnProposal(currentMarketData, id);
+  // const { data: powerAtProposalStart } = useVotingPowerAt(currentMarketData, strategy, startBlock);
 
-  const voteOngoing = state === ProposalState.Active;
+  const voteOnProposal = {
+    support: true
+  }
+
+  const powerAtProposalStart = 0
+
+  const voteOngoing = proposal.proposalData.proposalData.state === ProposalV3State.Active;
 
   // Messages
+  /*
   const didVote = powerAtProposalStart && voteOnProposal?.votingPower !== '0';
   const showAlreadyVotedMsg = !!user && voteOnProposal && didVote;
+  */
+  const didVote = true
+  const showAlreadyVotedMsg = !!user && didVote
+
   const showCannotVoteMsg = !!user && voteOngoing && Number(powerAtProposalStart) === 0;
   const showCanVoteMsg =
     powerAtProposalStart && !didVote && !!user && voteOngoing && Number(powerAtProposalStart) !== 0;
 
+
   return (
-    <>
+    <Paper sx={{ px: 6, py: 4, mb: 2.5 }}>
       <Typography variant="h3" sx={{ mb: 8 }}>
         <Trans>Your voting info</Trans>
       </Typography>
@@ -88,7 +105,7 @@ export function VoteInfo({ id, state, strategy, startBlock }: CustomProposalType
                 color="success"
                 variant="contained"
                 fullWidth
-                onClick={() => openGovVote(id, true, powerAtProposalStart)}
+                onClick={() => openGovVote(+proposal.proposal.proposalId, true, powerAtProposalStart)}
               >
                 <Trans>Vote YAE</Trans>
               </Button>
@@ -96,7 +113,7 @@ export function VoteInfo({ id, state, strategy, startBlock }: CustomProposalType
                 color="error"
                 variant="contained"
                 fullWidth
-                onClick={() => openGovVote(id, false, powerAtProposalStart)}
+                onClick={() => openGovVote(+proposal.proposal.proposalId, false, powerAtProposalStart)}
                 sx={{ mt: 2 }}
               >
                 <Trans>Vote NAY</Trans>
@@ -107,6 +124,6 @@ export function VoteInfo({ id, state, strategy, startBlock }: CustomProposalType
       ) : (
         <ConnectWalletButton />
       )}
-    </>
+    </Paper>
   );
 }
