+++ src/modules/governance/proposal/VotersList.tsx
@@ -2,12 +2,12 @@ import { Box, Theme } from '@mui/material';
 import { SxProps } from '@mui/system';
 import { Fragment } from 'react';
 
-import { GovernanceVoter } from './VotersListContainer';
 import { VotersListItem } from './VotersListItem';
+import { ProposalVote } from 'src/hooks/governance/useProposalVotes';
 
 type VotersListProps = {
   compact?: boolean;
-  voters: GovernanceVoter[];
+  voters: ProposalVote[];
   sx?: SxProps<Theme>;
 };
 
@@ -18,7 +18,7 @@ export const VotersList = ({ compact = false, voters, sx }: VotersListProps): JS
         <Box sx={{ color: 'text.secondary' }}>—</Box>
       ) : (
         voters.map((voter) => (
-          <Fragment key={voter.address}>
+          <Fragment key={voter.voter}>
             <VotersListItem voter={voter} compact={compact} />
           </Fragment>
         ))
