+++ src/modules/governance/proposal/VotersListContainer.tsx
@@ -1,29 +1,28 @@
 import { Trans } from '@lingui/macro';
 import { Box, Button, CircularProgress, Typography, useMediaQuery, useTheme } from '@mui/material';
-import fetch from 'isomorphic-unfetch';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Row } from 'src/components/primitives/Row';
 import { CustomProposalType } from 'src/static-build/proposal';
 import { useRootStore } from 'src/store/root';
 import { AIP } from 'src/utils/mixPanelEvents';
 
-import { formatProposal } from '../utils/formatProposal';
+import { FormattedProposalV3, formatProposal } from '../utils/formatProposal';
 import { VotersList } from './VotersList';
 import { VotersListModal } from './VotersListModal';
+import { ProposalVotes } from 'src/hooks/governance/useProposalVotes';
 
 type VotersListProps = {
-  proposal: CustomProposalType;
+  proposal: FormattedProposalV3;
+  proposalVotes: ProposalVotes;
 };
 
 export type GovernanceVoter = {
-  address: string; //"0x8298a996a00835eedcc75763038b731fe617fd0d"
-  ensName: string | null; //null
+  address: string;
+  ensName: string | null;
   votingPower: number;
   twitterAvatar: string | null;
   support: boolean;
 };
-
-type GovernanceProposalTopVotersResponse = [GovernanceVoter[], GovernanceVoter[]];
 
 export type VotersData = {
   yaes: GovernanceVoter[];
@@ -31,59 +30,18 @@ export type VotersData = {
   combined: GovernanceVoter[];
 };
 
-const sortByVotingPower = (a: GovernanceVoter, b: GovernanceVoter) => {
-  return a.votingPower < b.votingPower ? 1 : a.votingPower > b.votingPower ? -1 : 0;
-};
-
-export const VotersListContainer = (props: VotersListProps): JSX.Element => {
-  const { proposal } = props;
-  const { id: proposalId, forVotes, againstVotes } = proposal;
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<boolean>(false);
-  const [voters, setVoters] = useState<VotersData>();
+export const VotersListContainer = ({ proposal, proposalVotes }: VotersListProps): JSX.Element => {
   const [votersModalOpen, setVotersModalOpen] = useState(false);
   const { breakpoints } = useTheme();
   const mdScreen = useMediaQuery(breakpoints.only('md'));
-
-  const votersUrl = `${process.env.NEXT_PUBLIC_API_BASEURL}/data/proposal-votes`;
-  const queryParams = `?proposal=${proposalId}`;
   const trackEvent = useRootStore((store) => store.trackEvent);
-
-  const getVoterInfo = async () => {
-    if (error) setError(false);
-
-    try {
-      // Get proposal voters data
-      const resp = await fetch(votersUrl + queryParams);
-
-      if (resp.ok) {
-        const [yaes, nays]: GovernanceProposalTopVotersResponse = await resp.json();
-        const votersData: VotersData = {
-          yaes: yaes.sort(sortByVotingPower),
-          nays: nays.sort(sortByVotingPower),
-          combined: yaes.concat(nays).sort(sortByVotingPower),
-        };
-        setVoters(votersData);
-      } else {
-        setError(true);
-      }
-    } catch (e: unknown) {
-      console.error(e);
-      setError(true);
-    }
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    getVoterInfo();
-  }, [forVotes, againstVotes]); /* eslint-disable-line react-hooks/exhaustive-deps */
 
   const handleOpenAllVotes = () => {
     trackEvent(AIP.VIEW_ALL_VOTES);
     setVotersModalOpen(true);
   };
 
-  if (loading)
+  if (false)
     return (
       <Box sx={{ mt: 8, mb: 12 }}>
         <Row sx={{ mb: 3 }}>
@@ -97,7 +55,7 @@ export const VotersListContainer = (props: VotersListProps): JSX.Element => {
       </Box>
     );
 
-  if (error)
+  if (false)
     return (
       <Box sx={{ mt: 8, mb: 12 }}>
         <Row sx={{ mb: 3 }}>
@@ -113,13 +71,13 @@ export const VotersListContainer = (props: VotersListProps): JSX.Element => {
       </Box>
     );
 
-  if (!voters || voters.combined.length === 0) return <Box sx={{ mt: 8 }} />;
+  if (!proposalVotes || proposalVotes.combinedVotes.length === 0) return <Box sx={{ mt: 8 }} />;
 
   return (
     <Box sx={{ my: 8 }}>
       <Row sx={{ mb: 3 }}>
         <Typography variant="subheader2" color="text.secondary">
-          {voters.combined.length > 10 ? <Trans>Top 10 addresses</Trans> : <Trans>Addresses</Trans>}
+          {proposalVotes.combinedVotes.length > 10 ? <Trans>Top 10 addresses</Trans> : <Trans>Addresses</Trans>}
         </Typography>
         <Typography variant="subheader2" color="text.secondary">
           <Trans>Votes</Trans>
@@ -127,10 +85,10 @@ export const VotersListContainer = (props: VotersListProps): JSX.Element => {
       </Row>
       <VotersList
         compact={mdScreen}
-        voters={voters.combined.slice(0, 10)}
+        voters={proposalVotes.combinedVotes.slice(0, 10)}
         sx={{ my: 4, pr: 2.25 }}
       />
-      {voters.combined.length > 10 && (
+      {proposalVotes.combinedVotes.length > 10 && (
         <Button variant="outlined" fullWidth onClick={handleOpenAllVotes}>
           <Trans>View all votes</Trans>
         </Button>
@@ -139,8 +97,8 @@ export const VotersListContainer = (props: VotersListProps): JSX.Element => {
         <VotersListModal
           open={votersModalOpen}
           close={() => setVotersModalOpen(false)}
-          proposal={formatProposal(proposal)}
-          voters={voters}
+          proposal={proposal}
+          voters={proposalVotes}
         />
       )}
     </Box>
