+++ src/modules/governance/proposal/VotersListModal.tsx
@@ -6,16 +6,16 @@ import StyledToggleButton from 'src/components/StyledToggleButton';
 import StyledToggleButtonGroup from 'src/components/StyledToggleButtonGroup';
 
 import { BasicModal } from '../../../components/primitives/BasicModal';
-import { FormattedProposal } from '../utils/formatProposal';
+import { FormattedProposalV3 } from '../utils/formatProposal';
 import { VoteBar } from '../VoteBar';
 import { VotersList } from './VotersList';
-import { VotersData } from './VotersListContainer';
+import { ProposalVotes } from 'src/hooks/governance/useProposalVotes';
 
 type VotersListModalProps = {
   open: boolean;
   close: () => void;
-  proposal: FormattedProposal;
-  voters: VotersData;
+  proposal: FormattedProposalV3;
+  voters: ProposalVotes;
 };
 
 export const VotersListModal = ({
@@ -39,8 +39,8 @@ export const VotersListModal = ({
     <>
       <VoteBar
         yae
-        percent={proposal.yaePercent}
-        votes={proposal.yaeVotes}
+        percent={proposal.forPercent}
+        votes={proposal.forPercent}
         sx={{
           ...borderBaseStyle,
           px: 4,
@@ -57,14 +57,14 @@ export const VotersListModal = ({
           }}
         >
           <Typography variant="subheader2" color="text.secondary">
-            <Trans>Addresses ({voters.yaes.length})</Trans>
+            <Trans>Addresses ({voters.yaeVotes.length})</Trans>
           </Typography>
           <Typography variant="subheader2" color="text.secondary">
             <Trans>Votes</Trans>
           </Typography>
         </Row>
         <VotersList
-          voters={voters.yaes}
+          voters={voters.yaeVotes}
           sx={{
             p: 4,
             mb: 0,
@@ -78,8 +78,8 @@ export const VotersListModal = ({
   const noVotesUI = (
     <>
       <VoteBar
-        percent={proposal.nayPercent}
-        votes={proposal.nayVotes}
+        percent={proposal.againstPercent}
+        votes={proposal.againstPercent}
         sx={{
           ...borderBaseStyle,
           px: 4,
@@ -96,14 +96,14 @@ export const VotersListModal = ({
           }}
         >
           <Typography variant="subheader2" color="text.secondary">
-            <Trans>Addresses ({voters.nays.length})</Trans>
+            <Trans>Addresses ({voters.nayVotes.length})</Trans>
           </Typography>
           <Typography variant="subheader2" color="text.secondary">
             <Trans>Votes</Trans>
           </Typography>
         </Row>
         <VotersList
-          voters={voters.nays}
+          voters={voters.nayVotes}
           sx={{
             p: 4,
             mb: 0,
